refactor: extract conversation and read state
